+++ asset_allocation/strategies/momentum.py
@@ -74,13 +74,19 @@ class Momentum(strategy.Strategy):
             array of indexes
         """
         nan_sum = np.isnan(self.latest_return).sum()
-        top_n = min(self.top_n, self.num_total_assets)
-        missing_assets = (self.num_total_assets - nan_sum) - top_n
-        if missing_assets < 0:
-            return (-self.return_metrics_intuitive).argsort()[: self.top_n][
-                :missing_assets
-            ]
-        return (-self.return_metrics_intuitive).argsort()[: self.top_n]
+        top_n = min(self.top_n, self.num_total_assets - nan_sum)
+        neg_sort = (-self.return_metrics_intuitive).argsort()
+
+        selected_assets = 0
+        i = 0
+        a = list()
+
+        while selected_assets < top_n:
+            if self.index_comp[neg_sort[i]]:
+                a.append(neg_sort[i])
+                selected_assets += 1
+            i += 1
+        return np.array(a)
 
     @property
     def return_metrics_optimizer(self) -> np.array: